+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Rename `BEM` option on `SelectorFormat` to `strict_BEM` to emphasize that
   this is a particular dialect which may not be what developers expect
+* Allow `SelectorFormat` convention explanations to be customized via the
+  `#{type}_convention_explanation` and `convention_explanation` options
 
 ## 0.36.1